requirements.txt
tweepy
pandas
numpy
textblob
scikit-learn
transformers
torch
matplotlib
seaborn
wordcloud
plotly
python-dotenv
streamlit
